requirements.txt
Django==1.6.3
Unidecode==0.04.16
cssmin==0.2.0
django-assets==0.10
django-email-extras==0.3
django-forms-builder==0.11.1
eea.frame==0.3.2
future==0.9.0
gunicorn==19.1.1
jsmin==2.0.11
python-gnupg==0.3.6
requests==2.2.1
sphinx-me==0.3
webassets==0.10
wsgiref==0.1.2
